+++ Tachometer.h
@@ -8,7 +8,7 @@ class Tachometer
 {
   public:
     Tachometer(uint8_t spikesAllowed, uint8_t pulsesPerRevolution);
-    long getEngineSpeed();
+    uint16_t getEngineSpeed();
     void processInterrupt();
     
   private:
@@ -16,16 +16,16 @@ class Tachometer
     uint8_t _delimeter;
     
     //RPM calculation related
-    long _lastInterruptTime;                     //Last interruption time
-    long _rpm;                     //Current RPM value
-    long _prevRpm;                    //Previous RPM value
+    uint32_t _lastInterruptTime;                      //Last interruption time
+    uint16_t _rpm;                                    //Current RPM value
+    uint16_t _prevRpm;                                //Previous RPM value
     
     //Speed spikes catcher globals
     uint8_t _maxSpikes;
     uint8_t _spikesCounter;
     
     //Helper methods
-    bool isSpike(double rpm1, double rpm2);
+    bool isSpike(uint16_t rpm1, uint16_t rpm2);
     bool isRunning();
 
 };
